fix: configure frontend axios base url
