+++ frontend/.env.example
@@ -1,5 +1,9 @@
-# Change this to http://localhost:5000/api for local development
-VITE_API_URL=https://dailyforge-backend.onrender.com/api/
+# Copy this file to .env for local development.
+# The frontend uses this URL for all backend API requests.
+VITE_API_URL=http://localhost:5000/api
+
+# Production backend:
+# VITE_API_URL=https://dailyforge-backend.onrender.com/api
 
 # Firebase client configuration for Google Sign-In
 VITE_FIREBASE_API_KEY=your_firebase_api_key
