+++ frontend/src/api/axios.js
@@ -1,28 +1,1 @@
-import axios from "axios";
-
-// create axios instance
-const api = axios.create({
-  baseURL: import.meta.env.VITE_API_BASE_URL || import.meta.env.VITE_API_URL || (import.meta.env.DEV ? "http://localhost:5000/api/" : "https://dailyforge-backend.onrender.com/api/"),
-  timeout: Number(import.meta.env.VITE_API_TIMEOUT) || 15000, // updated 15s as default
-  withCredentials: true,
-});
-
-// Handle response errors, including timeout
-api.interceptors.response.use(
-  (response) => response, // success — pass through unchanged
-  (error) => {
-    if (error.code === "ECONNABORTED") {
-      // This fires when the timeout is hit
-      console.error("Request timed out. The server may be waking up from sleep. Please wait a moment and try again.");
-      error.userMessage =
-        "The server is waking up — this can take up to 30 seconds on first load. Please try again shortly.";
-    } else if (!error.response) {
-      // Network error (no internet, server completely unreachable)
-      console.error("Network error. Please check your connection.");
-      error.userMessage = "Network error. Please check your internet connection.";
-    }
-    return Promise.reject(error); // always reject so callers can handle it
-  }
-);
-
-export default api;
+export { default } from "./axiosConfig";
